Create Separate Thread for Audio Feedback
A thread for playing audio to the user is created to run in parallel
with the main thread (which is the Pomodoro application). the thread is
necessary since without it the main thread will be unresponsive until
the audio being played has finished.

The audio is be played once the Pomodoro timer hits 00:00.