Create Test for Pomodoro Function
In order to start testing, the pytest module will be used and the plugin
pytest-qt to simplify testing the PyQt GUI framework. The test target
is the app's text to string method wherein the test will assert if
the provided time in seconds will translate into a minutes and seconds
format (mm:ss).